chore: fix file mgmt stuff for tests
