+++ pkg/pipelinecontrol/pipelinecontrol.go
@@ -1,12 +1,9 @@
 package pipelinecontrol
 
 import (
-	"testing"
-
 	"github.com/newrelic/newrelic-client-go/v2/internal/http"
 	"github.com/newrelic/newrelic-client-go/v2/pkg/config"
 	"github.com/newrelic/newrelic-client-go/v2/pkg/logging"
-	mock "github.com/newrelic/newrelic-client-go/v2/pkg/testhelpers"
 )
 
 type Pipelinecontrol struct {
@@ -23,15 +20,4 @@ func New(config config.Config) Pipelinecontrol {
 		config: config,
 	}
 	return pkg
-}
-
-func newMockClient(t *testing.T, mockJSONResponse string, statusCode int) Pipelinecontrol {
-	ts := mock.NewMockServer(t, mockJSONResponse, statusCode)
-	tc := mock.NewTestConfig(t, ts)
-	return New(tc)
-}
-
-func newIntegrationTestClient(t *testing.T) Pipelinecontrol {
-	tc := mock.NewIntegrationTestConfig(t)
-	return New(tc)
 }
